Implement pick macro function
