+++ derive/src/utilities/mod.rs
@@ -1,13 +1,13 @@
-use darling::FromDeriveInput;
-use surreal_derive_helpers::utilities::TableDeriveAttributesPickable;
-use syn::parse_macro_input;
-use quote::quote;
-
-pub fn generate_table_resources_trait(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
-    let input = parse_macro_input!(input);
-    let output = match TableDeriveAttributesPickable::from_derive_input(&input) {
-        Ok(out) => out,
-        Err(err) => return proc_macro::TokenStream::from(err.write_errors()),
-    };
-    quote!(#output).into()
-}
+// pub mod pick;
+//
+// use syn;
+// use quote::quote;
+//
+// pub fn generate_pick_macro_fn_tokens(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
+//     let output = match syn::parse2::<PickedMeta>(input.into()) {
+//         Ok(out) => out,
+//         Err(err) => return err.to_compile_error().into()
+//     };
+//
+//     quote!(#output).into()
+// }
